[DP-2107] move athena code to the python base image (#2421)
* Move create_glue_database function into glue_utils

* Move CuratedDataQueryBuilder into base image

* Move CuratedDataLoader into python base

* Bump version

* move more potentially shared methods out of loader class into utils

* update tests

* remove commented out code

* lint

* did not mock athena

* one more lint

* the impports

* updated changelog

---------
