+++ containers/daap-athena-load/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Moved internal classes to base image (7.2.0)
+
 ## [1.5.0] 2023-11-15
 
 - Updated base image to 7.0.0
