+++ containers/daap-python-base/config.json
@@ -1,5 +1,5 @@
 {
   "name": "daap-python-base",
-  "version": "7.2.1",
+  "version": "7.3.0",
   "registry": "ghcr"
 }
